+++ src/utils/errors.ts
@@ -1,17 +1,23 @@
+import { ApiCallError } from 'iexec/errors';
 import { ValidationError } from 'yup';
 
-class WorkflowError extends Error {
-  originalError?: Error;
-
-  constructor(message: string, originalError?: Error) {
-    super(message);
+export class WorkflowError extends Error {
+  constructor(message: string, cause: Error) {
+    super(message, { cause: cause });
     this.name = this.constructor.name;
-    if (originalError) {
-      this.originalError = originalError;
-    }
   }
 }
 
-class NoValueError extends Error {}
+export function handleProtocolError(error: Error): boolean {
+  if (error instanceof ApiCallError) {
+    throw new WorkflowError(
+      "A service in the iExec protocol appears to be unavailable. You can retry later or contact iExec's technical support for help.",
+      error
+    );
+  }
+  return false;
+}
 
-export { ValidationError, WorkflowError, NoValueError };
+export class NoValueError extends Error {}
+
+export { ValidationError };